# src/main/resources/openapi/nayon-api-v1.yaml
openapi: 3.1.0
info:
  title: NYAON API
  version: 1.8.0
  description: Social-account-isolated profile, cloud save, economy, store purchase, subscription, reward, gacha, and battle audit API.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /subscriptions/catalog:
    get:
      operationId: getSubscriptionCatalog
      description: Returns independent monthly subscription products and server-owned benefit values. Prices remain owned by Google Play.
      parameters:
        - name: platform
          in: query
          required: true
          schema:
            type: string
            enum: [GOOGLE_PLAY]
      responses:
        '200':
          description: Active monthly subscription catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionCatalogResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /me/subscriptions:
    get:
      operationId: getMySubscriptions
      responses:
        '200':
          description: Current account subscription entitlement states
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /store/subscriptions/google-play/verify:
    post:
      operationId: verifyGooglePlaySubscription
      description: Verifies Google subscriptions v2 state and updates one independent account entitlement.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionVerifyRequest'
      responses:
        '200':
          description: Existing result for an identical request or token retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionVerifyResponse'
        '201':
          description: Subscription state verified and persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionVerifyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /me/level-rewards:
    get:
      operationId: getMyLevelRewards
      responses:
        '200':
          description: Server account level, independent subscription states, and lifetime claim state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelRewardListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /me/level-rewards/{trackCode}/{requiredLevel}/claim:
    post:
      operationId: claimLevelReward
      parameters:
        - name: trackCode
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/LevelRewardTrackCode'
        - name: requiredLevel
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 50
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Existing result for an identical retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelRewardClaimResponse'
        '201':
          description: Lifetime reward granted once by the authoritative ledger
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelRewardClaimResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    googleOidc:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID generated once for a logical write command.
      schema:
        type: string
        format: uuid
      example: 6bf5aa51-f26c-42df-bfc4-f52c4cae1356
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Revision, idempotency, duplicate command, or invalid current-state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnprocessableEntity:
      description: Valid command cannot be completed with the current account state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    TooManyRequests:
      description: Request rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ServiceUnavailable:
      description: Dependent service is unavailable or not configured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    SubscriptionPlanCode:
      type: string
      enum: [MONTHLY_GROWTH, MONTHLY_ADVANCED]
    SubscriptionState:
      type: string
      enum: [PENDING, ACTIVE, CANCELED, GRACE_PERIOD, ON_HOLD, PAUSED, EXPIRED, REVOKED]
    SubscriptionBenefit:
      type: object
      additionalProperties: false
      required: [code, value, version]
      properties:
        code:
          type: string
          enum: [INITIAL_DIAMOND, DAILY_DIAMOND, SKILL_REFRESH_COUNT, REVIVE_COUNT, MAX_ENERGY, GOLD_BONUS_BPS, BATTLE_SPEED_UNLOCK, REWARDED_AD_SKIP]
        value:
          type: integer
          format: int64
          minimum: 1
        version:
          type: integer
          minimum: 1
    SubscriptionCatalogPlan:
      type: object
      additionalProperties: false
      required: [planCode, productId, productType, rewardTrackCode, benefits]
      properties:
        planCode:
          $ref: '#/components/schemas/SubscriptionPlanCode'
        productId:
          type: string
          minLength: 1
          maxLength: 200
        productType:
          type: string
          enum: [SUBSCRIPTION]
        rewardTrackCode:
          $ref: '#/components/schemas/LevelRewardTrackCode'
        benefits:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionBenefit'
    SubscriptionCatalogResponse:
      type: object
      additionalProperties: false
      required: [platform, obfuscatedAccountId, plans]
      properties:
        platform:
          type: string
          enum: [GOOGLE_PLAY]
        obfuscatedAccountId:
          type: string
          pattern: '^[0-9a-f]{64}$'
        plans:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionCatalogPlan'
    SubscriptionEntitlement:
      type: object
      additionalProperties: false
      required: [planCode, state, entitled, autoRenewing]
      properties:
        planCode:
          $ref: '#/components/schemas/SubscriptionPlanCode'
        state:
          $ref: '#/components/schemas/SubscriptionState'
        entitled:
          type: boolean
        autoRenewing:
          type: boolean
        startedAt:
          type: [string, 'null']
          format: date-time
        expiresAt:
          type: [string, 'null']
          format: date-time
        lastVerifiedAt:
          type: [string, 'null']
          format: date-time
    SubscriptionListResponse:
      type: object
      additionalProperties: false
      required: [serverTime, subscriptions]
      properties:
        serverTime:
          type: string
          format: date-time
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionEntitlement'
    SubscriptionVerifyRequest:
      type: object
      additionalProperties: false
      required: [productId, purchaseToken]
      properties:
        productId:
          type: string
          minLength: 1
          maxLength: 200
        purchaseToken:
          type: string
          minLength: 1
          maxLength: 4096
    SubscriptionRewardGrant:
      type: object
      additionalProperties: false
      required: [assetCode, amount, balance]
      properties:
        assetCode:
          type: string
          enum: [DIAMOND]
        amount:
          type: integer
          format: int64
          minimum: 1
        balance:
          type: integer
          format: int64
          minimum: 0
    SubscriptionVerifyResponse:
      type: object
      additionalProperties: false
      required: [subscription, replay]
      properties:
        subscription:
          $ref: '#/components/schemas/SubscriptionEntitlement'
        initialReward:
          oneOf:
            - $ref: '#/components/schemas/SubscriptionRewardGrant'
            - type: 'null'
        replay:
          type: boolean
...
    LevelRewardTrackCode:
      type: string
      enum: [FREE, PREMIUM, ROYAL]
    LevelRewardCatalogItem:
      type: object
      additionalProperties: false
      required: [version, trackCode, requiredLevel, assetType, assetCode, amount, claimed, claimable]
      properties:
        version:
          type: integer
          minimum: 1
        trackCode:
          $ref: '#/components/schemas/LevelRewardTrackCode'
        requiredLevel:
          type: integer
          minimum: 1
          maximum: 50
        assetType:
          type: string
          enum: [CURRENCY, ITEM]
        assetCode:
          type: string
          minLength: 1
          maxLength: 80
        amount:
          type: integer
          format: int64
          minimum: 1
        claimed:
          type: boolean
        claimable:
          type: boolean
    LevelRewardListResponse:
      type: object
      additionalProperties: false
      required: [accountLevel, totalAccountExp, subscriptions, rewards]
      properties:
        accountLevel:
          type: integer
          minimum: 1
          maximum: 50
        totalAccountExp:
          type: integer
          format: int64
          minimum: 0
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionEntitlement'
        rewards:
          type: array
          items:
            $ref: '#/components/schemas/LevelRewardCatalogItem'
    LevelRewardClaimResponse:
      type: object
      additionalProperties: false
      required: [claimId, reward, economy, replay]
      properties:
        claimId:
          type: string
          format: uuid
        reward:
          $ref: '#/components/schemas/LevelRewardCatalogItem'
        economy:
          $ref: '#/components/schemas/EconomyResponse'
        replay:
          type: boolean
...
    EconomyResponse:
      type: object
      additionalProperties: false
      required: [bootstrapped, currencies, items, equipment]
      properties:
        bootstrapped:
          type: boolean
        currencies:
          type: object
          additionalProperties:
            type: integer
            format: int64
            minimum: 0
        items:
          type: object
          additionalProperties:
            type: integer
            format: int64
            minimum: 0
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/PlayerEquipment'
    PlayerEquipment:
      type: object
      additionalProperties: false
      required: [id, equipmentCode, grade, level, locked]
      properties:
        id:
          type: string
          format: uuid
        equipmentCode:
          type: string
          maxLength: 80
        grade:
          type: string
          enum: [COMMON, UNCOMMON, RARE, EPIC, UNIQUE]
        level:
          type: integer
          minimum: 1
        locked:
          type: boolean
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, traceId, details]
      properties:
        code:
          type: string
          example: SAVE_REVISION_CONFLICT
        message:
          type: string
          example: The cloud save has changed.
        traceId:
          type: string
        details:
          type: object
          additionalProperties: true